requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
pytz==2024.1
typing_extensions==4.12.1
yarl==1.9.4
